cores/requirements.txt
astroid==2.2.5
Click==7.0
Django==2.2.2
django-bootstrap4==0.0.8
django-filter==2.1.0
django-js-asset==1.2.2
django-mptt==0.10.0
django-netfields==0.10.0
django-polymorphic==2.0.3
djangorestframework==3.9.4
isort==4.3.20
Jinja2==2.10.1
lazy-object-proxy==1.4.1
livereload==2.6.1
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
netaddr==0.7.19
psycopg2-binary==2.8.2
pylint==2.3.1
pytz==2019.1
PyYAML==5.1.1
six==1.12.0
sqlparse==0.3.0
tornado==6.0.2
typed-ast==1.4.0
wrapt==1.11.1
uWSGI==2.0.18
django-auth-ldap==2.0.0
django-auth-kerberos==1.2.5
pykerberos==1.2.1
coreapi==2.3.3
coreschema==0.0.4
itypes==1.1.0
uritemplate==3.0.0
django-impersonate==1.4.1